Decluttered the Main Client library class by separating transactions into their own classes based on payment method.
