+++ pos-client-library/src/main/java/com/globalaccelerex/globalaccelerexandroidposclientlibrary/baseAppUtils/TransactionRequest.kt
@@ -7,25 +7,27 @@ import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.B
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.MOBILE_MONEY_REQUEST_DATA_TAG
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.MOBILE_MONEY_TRANSACTION
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.MOBILE_MONEY_TRANSACTION_INTENT_ADDRESS
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_PURCHASE
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_PURCHASE_WITH_CASH_BACK
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.REQUEST_DATA_TAG
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_REQUEST_INTENT_ADDRESS
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_CARD_BALANCE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_PRE_AUTH_PURCHASE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_PURCHASE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_PURCHASE_WITH_CASH_BACK
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_REFUND
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.baseAppUtils.BaseAppConstants.TRANSACTION_TYPE_REVERSAL
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.*
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_PURCHASE_REQUEST_CODE
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_PURCHASE_WITH_CB_REQUEST_CODE
-import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.MOBILE_MONEY_STATUS_CHECK_REQUEST_CODE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.exceptions.UnsupportedCallingComponentException
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CARD_PURCHASE_REQUEST_CODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CARD_PURCHASE_WITH_CB_REQUEST_CODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_CARD_BALANCE_REQUEST_CODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_PRE_AUTH_COMPLETION_REQUEST_CODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_PRE_AUTH_PURCHASE_REQUEST_CODE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_PURCHASE_REQUEST_CODE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_PURCHASE_WITH_CB_REQUEST_CODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.CNP_REFUND_REQUEST_CODE
 import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.MOBILE_MONEY_PURCHASE_REQUEST_CODE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.GaResponseKeys.MOBILE_MONEY_STATUS_CHECK_REQUEST_CODE
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.MobileMoneyOperators
+import com.globalaccelerex.globalaccelerexandroidposclientlibrary.util.toPosAmount
 import com.google.gson.Gson
 
 internal class TransactionRequest {
@@ -44,10 +46,16 @@ internal class TransactionRequest {
         val transJson = Gson().toJson(transactionObject)
         val intent = Intent(TRANSACTION_REQUEST_INTENT_ADDRESS)
         intent.putExtra(REQUEST_DATA_TAG, transJson)
-        if (callingComponent is Fragment) {
-            callingComponent.startActivityForResult(intent, CARD_PURCHASE_REQUEST_CODE)
-        } else if (callingComponent is Activity) {
-            callingComponent.startActivityForResult(intent, CARD_PURCHASE_REQUEST_CODE)
+        when (callingComponent) {
+            is Fragment -> {
+                callingComponent.startActivityForResult(intent, CARD_PURCHASE_REQUEST_CODE)
+            }
+            is Activity -> {
+                callingComponent.startActivityForResult(intent, CARD_PURCHASE_REQUEST_CODE)
+            }
+            else -> {
+                throw UnsupportedCallingComponentException("Unsupported calling component.")
+            }
         }
     }
