+++ pos-client-library/src/main/java/com/globalaccelerex/globalaccelerexandroidposclientlibrary/exceptions/ClientAppExceptions.kt
@@ -6,3 +6,8 @@ class UnsupportedFeatureException(private val customMessage: String): Exception(
     override val message: String?
         get() = this.customMessage
 }
+
+class UnsupportedCallingComponentException(private val customMessage: String): Exception() {
+    override val message: String?
+        get() = this.customMessage
+}
